+++ src/libtcod/sdl2/gl_alias.cpp
@@ -56,98 +56,206 @@ class AliasPool {
   std::mutex mutex_;
   std::map<Key, std::weak_ptr<T>> alias_pool_;
 };
-class OpenGLInternalTilesetAlias_ : public TilesetObserver {
+class OpenGLTilesetAlias::impl : public TilesetObserver {
+  using pool_type = AliasPool<
+      impl, const Tileset*, const std::shared_ptr<Tileset>&>;
  public:
-  OpenGLInternalTilesetAlias_(const std::shared_ptr<Tileset>& tileset)
+  impl(const std::shared_ptr<Tileset>& tileset)
   : TilesetObserver(tileset)
   {
-    SDL_AddEventWatch(on_sdl_event, this);
     glGenTextures(1, &gltexture_); gl_check();
-    sync_alias();
+    clear_alias();
   }
-  ~OpenGLInternalTilesetAlias_() {
-    SDL_DelEventWatch(on_sdl_event, this);
-    if (gltexture_) {
-      glDeleteTextures(1, &gltexture_); gl_check();
+  auto prepare_alias(const TCOD_Console& console) -> uint32_t
+  {
+    for (int i = 0; i < console.w * console.h; ++i) {
+      int codepoint = console.ch_array[i];
+      if (codepoint >= static_cast<int>(local_map_.size())
+          || local_map_.at(codepoint) < 0) {
+        if (ensure_tile(codepoint) == -1) {
+          // The texture alias was replaced by a bigger one.
+          i = -1; // Need to start over from the beginning.
+          continue;
+        }
+      }
     }
+    return gltexture_;
   }
-  GLuint gltexture_ = 0;
+  auto get_tile_position(int codepoint) const -> std::array<int, 2>
+  {
+    int index = local_map_.at(codepoint);
+    return {index % columns(), index / columns()};
+  }
+  auto get_alias_shape() const -> std::array<float, 2>
+  {
+    return {
+        static_cast<float>(texture_size_) / tile_width(),
+        static_cast<float>(texture_size_) / tile_height(),
+    };
+  }
+  auto get_alias_size() const -> std::array<float, 2>
+  {
+    return {
+        static_cast<float>(columns() * tile_width()),
+        static_cast<float>(rows() * tile_height()),
+    };
+  }
+  static pool_type pool_;
  protected:
   virtual void on_tileset_changed(
       const std::vector<std::pair<int, Tile&>> &changes) override
   {
-    sync_alias();
-  }
- private:
-  void sync_alias()
-  {
-    const std::vector<Tile>& tiles = tileset_->get_tiles();
-    int tile_width = tileset_->get_tile_width();
-    int tile_height = tileset_->get_tile_height();
-    int width = static_cast<int>(tile_width * tiles.size());
-    int height = tile_height;
-    Image alias(width, height);
-    for (int i = 0; i < tiles.size(); ++i) {
-      Image tile = tiles.at(i).get_image();
-      for (int y = 0; y < tile_height; ++y) {
-        for (int x = 0; x < tile_width; ++x) {
-          alias.at(x + tile_width * i, y) = tile.at(x, y);
-        }
+    for (const auto& changed : changes) {
+      // If this glyph is already in this alias then free it.
+      int codepoint = changed.second.codepoint;
+      if (codepoint < static_cast<int>(local_map_.size())
+          && local_map_.at(codepoint) >= 0) {
+        unallocated_.emplace_back(local_map_.at(codepoint));
+        local_map_.at(codepoint) = -1;
       }
     }
+  }
+ private:
+  /**
+   *  Return the rectangle for the tile `index` on the texture.
+   *
+   *  Does not perform any extra checks.
+   */
+  auto get_alias_rect(int index) const noexcept -> SDL_Rect
+  {
+    return SDL_Rect{
+        (index % columns()) * tile_width(),
+        (index / columns()) * tile_height(),
+        tile_width(),
+        tile_height(),
+    };
+  }
+  /**
+   *  Make sure a codepoint exists on this alias.
+   *
+   *  Returns 0 on success, or returns -1 if the texture was reset and
+   *  everything needs to be regenerated.
+   */
+  auto ensure_tile(int codepoint) -> int
+  {
+    if (codepoint >= static_cast<int>(local_map_.size())) {
+      local_map_.resize(codepoint + 1, -1);
+    }
+    if (next_alias_index_ == capacity()) {
+      // Replace the current texture with a larger one.
+      texture_size_ *= 2;
+      clear_alias();
+      return -1; // Tell prepare_alias to start over.
+    }
+    // Allocate an area of the texture alias for this codepoint.
+    SDL_Rect alias_rect;
+    if (unallocated_.size()) {
+      alias_rect = get_alias_rect(unallocated_.back());
+      unallocated_.pop_back();
+    } else {
+      alias_rect = get_alias_rect(next_alias_index_);
+      local_map_.at(codepoint) = next_alias_index_++;
+    }
+    // Find and draw the tile glyph into the alias.
+    const auto& charmap = tileset_->get_character_map();
+    Image tile;
+    if (codepoint >= static_cast<int>(charmap.size())
+        || charmap.at(codepoint) < 0) {
+      // Glyph not in this tileset, leave blank.
+      tile = Image(alias_rect.w, alias_rect.h, {0, 0, 0, 0});
+    } else{
+      tile = tileset_->get_tiles().at(charmap.at(codepoint)).get_image();
+    }
+    glBindTexture(GL_TEXTURE_2D, gltexture_); gl_check();
+    glTexSubImage2D(
+        GL_TEXTURE_2D,
+        0,
+        alias_rect.x,
+        alias_rect.y,
+        alias_rect.w,
+        alias_rect.h,
+        GL_RGBA,
+        GL_UNSIGNED_BYTE,
+        tile.data());
+    return 0; // Tile is ready.
+  }
+  /**
+   *  Replace the texture alias with an undefined texture.
+   *
+   *  This function depends on the value of `texture_size_` before calling.
+   */
+  void clear_alias()
+  {
+    for (auto& it : local_map_) { it = -1; }
+    unallocated_.clear();
+    next_alias_index_ = 0;
     glBindTexture(GL_TEXTURE_2D, gltexture_); gl_check();
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
     gl_check();
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA, width, height, 0, GL_RGBA,
-                 GL_UNSIGNED_BYTE, alias.data()); gl_check();
+    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA, texture_size_, texture_size_, 0,
+                 GL_RGBA, GL_UNSIGNED_BYTE, nullptr); gl_check();
+    glBindTexture(GL_TEXTURE_2D, 0); gl_check();
   }
-  static int on_sdl_event(void* userdata, SDL_Event* event)
-  {
-    auto this_ = static_cast<OpenGLInternalTilesetAlias_*>(userdata);
-    switch (event->type) {
-      case SDL_RENDER_DEVICE_RESET: // Lost all textures.
-        this_->sync_alias();
-        break;
-    }
-    return 0;
-  }
+  /**
+   *  Return the tile width of the assigned tileset.
+   */
+  int tile_width() const noexcept { return tileset_->get_tile_width(); }
+  /**
+   *  Return the tile height of the assigned tileset.
+   */
+  int tile_height() const noexcept { return tileset_->get_tile_height(); }
+  /**
+   *  Return the maximum columns supported by the current texture alias.
+   */
+  int columns() const noexcept { return texture_size_ / tile_width(); }
+  /**
+   *  Return the maximum rows supported by the current texture alias.
+   */
+  int rows() const noexcept { return texture_size_ / tile_height(); }
+  /**
+   *  Return the total number of tiles supported by the current texture alias.
+   */
+  int capacity() const noexcept { return columns() * rows(); }
+  int texture_size_ = 256;
+  GLuint gltexture_ = 0;
+  /**
+   *  The next index on the alias. Check `capacity` before using.
+   */
+  int next_alias_index_ = 0;
+  /**
+   *  Mapping of `Unicode -> texture_index`.
+   */
+  std::vector<int> local_map_;
+  /**
+   *  Alias indexes that have been temporarily freed by changed tiles.
+   */
+  std::vector<int> unallocated_;
 };
-static AliasPool<OpenGLInternalTilesetAlias_, const Tileset*,
-                 const std::shared_ptr<Tileset>&> opengl_alias_pool_;
+typename OpenGLTilesetAlias::impl::pool_type OpenGLTilesetAlias::impl::pool_;
+
 OpenGLTilesetAlias::OpenGLTilesetAlias(std::shared_ptr<Tileset> tileset)
-: alias_(opengl_alias_pool_.get(tileset.get(), tileset))
+: impl_(OpenGLTilesetAlias::impl::pool_.get(tileset.get(), tileset))
 {}
 const std::shared_ptr<Tileset>& OpenGLTilesetAlias::get_tileset()
 {
-  return alias_->get_tileset();
+  return impl_->get_tileset();
 }
-uint32_t OpenGLTilesetAlias::get_alias_texture()
+uint32_t OpenGLTilesetAlias::get_alias_texture(const TCOD_Console& console)
 {
-  return alias_->gltexture_;
+  return impl_->prepare_alias(console);
 }
-std::array<int, 2> OpenGLTilesetAlias::get_alias_shape()
+auto OpenGLTilesetAlias::get_alias_shape() const -> std::array<float, 2>
 {
-  return {
-      static_cast<int>(alias_->get_tileset()->get_tiles().size()),
-      1,
-  };
+  return impl_->get_alias_shape();
 }
-std::array<int, 2> OpenGLTilesetAlias::get_alias_size()
+auto OpenGLTilesetAlias::get_alias_size() const -> std::array<float, 2>
 {
-  auto shape = get_alias_shape();
-  return {
-      shape[0] * get_tileset()->get_tile_width(),
-      shape[1] * get_tileset()->get_tile_height(),
-  };
+  return impl_->get_alias_size();
 }
-std::array<int, 2> OpenGLTilesetAlias::get_tile_position(int codepoint)
+auto OpenGLTilesetAlias::get_tile_position(int codepoint) const
+-> std::array<int, 2>
 {
-  auto& charmap = alias_->get_tileset()->get_character_map();
-  int tile_index = 0;
-  if (static_cast<unsigned int>(codepoint) < charmap.size()) {
-    tile_index = charmap.at(codepoint);
-  }
-  return {tile_index, 1};
+  return impl_->get_tile_position(codepoint);
 }
 } // namespace sdl2
 } // namespace tcod